requirements.txt
streamlit
pandas
altair
matplotlib.pyplot
networkx
numpy
sklearn.ensemble
sklearn.linear_model
sklearn.model_selection
sklearn.metrics

